fix: convert ESLint configs to CommonJS for pre-commit compatibility
- Rename eslint.config.mjs to eslint.config.cjs for both projects
- Convert ES module imports to CommonJS require() syntax
- Update pre-commit config to use .cjs files
- Add .cjs to prettier file patterns
- Resolves ERR_MODULE_NOT_FOUND errors in pre-commit environment
- TypeScript ESLint now works properly with pre-commit hooks
