+++ src/cook-web/eslint.config.cjs
@@ -1,0 +1,49 @@
+const typescriptEslint = require('@typescript-eslint/eslint-plugin');
+const typescriptParser = require('@typescript-eslint/parser');
+
+const eslintConfig = [
+  {
+    files: ['**/*.{js,jsx,ts,tsx}'],
+    plugins: {
+      '@typescript-eslint': typescriptEslint,
+    },
+    languageOptions: {
+      parser: typescriptParser,
+      ecmaVersion: 'latest',
+      sourceType: 'module',
+      parserOptions: {
+        ecmaFeatures: {
+          jsx: true,
+        },
+      },
+    },
+    rules: {
+      // Basic JavaScript/TypeScript rules
+      'no-unused-vars': 'off',
+      '@typescript-eslint/no-unused-vars': 'error',
+      '@typescript-eslint/no-explicit-any': 'off',
+      'no-console': 'warn',
+
+      // React/Next.js specific rules
+      'react/react-in-jsx-scope': 'off', // Not needed with React 17+
+      'react/prop-types': 'off', // Using TypeScript for prop validation
+
+      // Next.js specific
+      '@next/next/no-img-element': 'off',
+      '@next/next/no-html-link-for-pages': 'off',
+    },
+  },
+  {
+    files: [
+      '**/*.test.{js,jsx,ts,tsx}',
+      '**/__tests__/**/*',
+      '**/scripts/**/*.js',
+      '**/jest.setup.js',
+    ],
+    rules: {
+      'no-console': 'off', // Allow console in tests, scripts, and test setup
+    },
+  },
+];
+
+module.exports = eslintConfig;
