--- src/web/eslint.config.mjs
@@ -1,43 +1,0 @@
-import typescriptEslint from "@typescript-eslint/eslint-plugin";
-import typescriptParser from "@typescript-eslint/parser";
-
-const eslintConfig = [
-  {
-    files: ["**/*.{js,jsx,ts,tsx}"],
-    plugins: {
-      "@typescript-eslint": typescriptEslint,
-    },
-    languageOptions: {
-      parser: typescriptParser,
-      ecmaVersion: "latest",
-      sourceType: "module",
-      parserOptions: {
-        ecmaFeatures: {
-          jsx: true,
-        },
-      },
-    },
-    rules: {
-      // Basic JavaScript/TypeScript rules
-      "no-unused-vars": "off",
-      "@typescript-eslint/no-unused-vars": "error",
-      "no-console": "warn",
-
-      // React specific rules
-      "react/react-in-jsx-scope": "off", // Not needed with React 17+
-      "react/prop-types": "off", // Using TypeScript for prop validation
-    },
-  },
-  {
-    files: [
-      "**/*.test.{js,jsx,ts,tsx}",
-      "**/__tests__/**/*",
-      "**/scripts/**/*.js",
-    ],
-    rules: {
-      "no-console": "off", // Allow console in tests and scripts
-    },
-  },
-];
-
-export default eslintConfig;
